refactor(core): move address derivation to standalone module to avoid SwapModule import chain
